Added a new option to $obvius->search: override_repeatable. This can be used to override the default behavior for repeatable fields which is to only replace the first occurance of the field in the search query. The default behavior has been made so you can search for two different values in a repeatable field at the same time by adding the same field to the fieldlist multiple times, eg. by doing a:
$obvius->(['rfield', 'rfield'], "rfield = 'val1' and rfield = 'val2'");

this allows you to search for documents which have both a value of 'val1'
and a value of 'val2' stored in "rfield".
The problem with the default behviour is that it isn't very effective. You
will have to join the whole vfields table for each time you want to search
for something in the rfield. Sometimes you know that you only need one
rfield to match your whole search expression against, eg. if you want to
do something like this:

$where = "category LIKE "01 CAT1%" OR category LIKE "02 CAT2%";

When you are sure of this you can set the override_repeatable option
for the field to make the search faster:

my %options;
$options{override_repeatable} = { category => 1 };

$obvius->search(['category'], $where, %options);