Added a few options from the config file that makes dealing with htdig easier:
htdig_sitename: This should be set to whatever hostname have been used
to create the htdig index so the correct "restict" options will be used
when calling htsearch.
report_htdig_running: if this config option is set to true the
"RUNNING "... debug lines will be printed with $obvius->log-error instead
of $obvius->log->debug. This way they will always show up in the error_log,
even though the site not running in debug mode. This is usefull for testing
htdig setups.