Bugfix: The regexp that changes fieldnames in the $where-string now only matches fieldnames that do not have a preceeding underscore, '_'.
This means that if you do a ComboSearch with, for instance, an
id-clause, like:

 id !1 3342

(which means that the document id must not be in the list (3342)) -
then Obvius::(MultiSite::)->search won't try to look in the
nonsensical and syntactically invalid field
'uri_cache.path_part1_obvius_documents.id', but rather in the expected
'uri_cache.path_part1_id'.

I don't think valid fieldnames ever come after an underscore in
WHERE-strings, so this change should have no impact on previously
working calls to Obvius::(MultiSite::)->search, as far as I can
see, and guess.